Move JwtVerifier.java to backendServices
