+++ finish/backendServices/src/test/java/it/io/openliberty/guides/jwt/SystemEndpointTest.java
@@ -21,7 +21,7 @@ import javax.ws.rs.core.Response.Status;
 import org.junit.Before;
 import org.junit.Test;
 import it.io.openliberty.guides.jwt.util.TestUtils;
-import test.JwtVerifier;
+import it.io.openliberty.guides.jwt.util.JwtVerifier;
 
 public class SystemEndpointTest {
